refactor(masonry): moved from isotope to masonry, and to my own variant of masonry: fluid-masonry
BREAKING CHANGE: styling of the masonry tile different.
the styling of the masonry tiles is a little bit different: we do not need a padding around each tile;
this is handled by the gutter option. we need a margin-bottom for every tile in the size of the gutter.

the gutter width and the minColumnWidt can be given by element sizing, means we can specify it in the css.